configmgr: Fix incorrect IdleCheck implementation + memleak
Commit d62cdb6af3ada85 did only a partially functional IdleCheck
implementation.  The IdleCheck never worked properly when a
configuration profile had been imported and then later on deleted; the
reference counting was never decreased.

This was due to two bugs.  The IdleCheck object pointer was not properly
passed to ConfigurationObjection.  In addition, the ConfigurationObject
was never truly deleted from memory when the Remove() D-Bus method was
called.

This fix also ensure the usage timestamp also gets updated when various
D-Bus calls occurs, regardless if the configuration manager keeps any
configuration profiles in memory.

Signed-off-by: David Sommerseth <davids@openvpn.net>