configmgr: Use ConfigManagerSignals class via inheritance
Just minor coding clean-up and simpliciations, which makes the coding
style match what is used in sessionmgr as well.  Also made it implicit
that LogGroup will always be CONFIGMGR when using the
ConfigManagerSignals class.

Also sneaked in a minor cleanup in the ConfigurationObject class and
ensure all declared class variables are initialized.

Signed-off-by: David Sommerseth <davids@openvpn.net>