Implement gathering of VPN tunnel statistics
This adds a 'statistics' object property exposed on both the VPN backend
client service and the session object provided by the session manager.
This property contains a dictionary of key/value pairs where the value
is always an int64 value.  The key field is a string containing a
description of the statistic value.

Added --stats operation to the openvpn3 command line interface to query
statistics for a specific session.  This interface will also dump the
statistics when disconnecting a session.

Also added a unit test program which dumps the statistics of an on going
VPN session to stdout; essentially a minified implementation of what is
present in openvpn3.

Signed-off-by: David Sommerseth <davids@openvpn.net>