backendstarter: Get rid of "Manager" references
The BackendManger approach was remodelled into a "starter service" only
quite earily in the development process.  But some of the old naming was
still lingering.  So all "manager" references are changed to "starter"
instead, which corresponds better with the reality.

Also re-implemented the signal handling to be a more core part of the
BackendStarter class.  This removes the need to separately track the
signal object.

Signed-off-by: David Sommerseth <davids@openvpn.net>