cli: Add openvpn2 command line interface
This interface tries to replicate the classing OpenVPN 2.x behaviour,
while essentially being just a D-Bus client for the OpenVPN 3 services.

Since this implementation is done in Python, a constants Enum generator
is added in addition.  This generator dumps out to stdout generated
StatusMajor and StatusMinor Python classes, based on the values defined
in dbus/constants.hpp.  Nothing fancy, but does the job.

Signed-off-by: David Sommerseth <davids@openvpn.net>