sessionmgr: Save StatusMinor::CONN_AUTH_FAILED situations
If the remote server sends a AUTH_FAILED message, the client should
disconnect and try again later with new credentials.  To avoid a
front-end user not to see this AUTH_FAILED, the session object
will not change the last_status after a AUTH_FAILED signal have
been caught.

Also implement logic so the session object in these AUTH_FAILED
scenarios are not destroyed but accessible to retrieve this last
status signal.

Signed-off-by: David Sommerseth <davids@openvpn.net>