configmgr: Don't count non-root Fetch calls as use
Only remove the configuration from memory when it is plausible it
has been retrieved by the openvpn3-service-client process (running as
root). This allows at least the configuration object owner to have a
look at the the configuration.  Before this change, just looking at the
configuration would also remove it from memory.

Signed-off-by: David Sommerseth <davids@openvpn.net>