cli: Add a new and more generic command line argument parser
This is built around the getopt_long() function but provides a more
generic C++ interface which automates most of the parsing logic.

It provides a simpler interface to collect a set of commands inside a
single program, where each program have their own set of options and
arguments.

The command line argument parsing (based on main()'s argc and argv) is
handled by this implementation and the result is sent to the commands
designated callback function via a parsed argument object.  This ensures
all the callback functions have the same command line processing plus
having a simpler API to extract the information it needs to complete its
job.

Signed-off-by: David Sommerseth <davids@openvpn.net>