proxy/configmgr: Added methods for managing public-access flags
Added SetPublicAccess() and GetPublicAccess() methods which is used to
manage the public_access property in configuration profiles.

When the public_access property is set to true, all users are allowed to
read and use the configuration profile - regardless of what the ACL
contains.  Only read access is granted with public_access.

If a configuration profile is also locked down, users can only use this
configuration profile to establish new VPN connections.  In this case
they cannot retrieve the contents of the configuration file.

Signed-off-by: David Sommerseth <davids@openvpn.net>