cli: Introduce get_userid() lookup function
This is just a simple helper function which takes a string containing
either a username or a UID value (as string) and returns the proper
uid_t value for the input.  If a username is not found, -1 is returned.

Updated get-acl test program to make use of this feature as well.

Signed-off-by: David Sommerseth <davids@openvpn.net>