cli/cmdparser: Split up SingleCommand::RunCommand() method
Split this method into a parsing and an execution step.  The parsing
step is now placed into a protected SingleCommand::parse_commandline()
method.  This enables the SingleCommand class to be inherited and the
RunCommand() method can be overridden without much code duplication.

This also required the ParsedArgs class to get an indicator which can
provide a feedback if the command line parsing completed or was
interrupted (like via -h/--help).

Signed-off-by: David Sommerseth <davids@openvpn.net>