cli/cmdparser: Implement support for generating shell completion hints
This extends the Commands class to be able to automatically generate
hints to calling shells containing values of possible openvpn3 commands,
options and even option arguments.

The list of commands and options for each command happens completely
automatically.  But to be able to produce related argument hints to
options, a callback function is needed to be implemented.

With this added, a new command is now added automatically,
shell-completion.  This command takes care of everything shell
completion scripts needs to provide proper command line completion.

The tests/command-parse/cmdparser.cpp test program has been updated to
provide possibilities to test the new shell-completion command.

Signed-off-by: David Sommerseth <davids@openvpn.net>