cmdargparser: Improve the struct option generation
The SingleCommand and SingleCommandOption classes generates struct
option on-the-fly according to the AddOption() calls.  This is to be
able to provide a more dynamically generated command line parser.

The initial implementation had some flaws, which got exposed on some
platforms.  This patch simplifies the pointer handling and tries to do
less magic with pointers.  In addition, it adds a safeguard for the name
member in struct option to ensure its content is always available.

Signed-off-by: David Sommerseth <davids@openvpn.net>