Improve version number presentation
Improved how the client presents the version number in particulary the
peer-info data sent to the server.  Now the git branch/commit reference
is somewhat easier to read out plus added the git flags identifying if
the client was built from a "dirty" git tree.

Also switched from using '/' as the branch/commit separator to ':'.

Signed-off-by: David Sommerseth <davids@openvpn.net>