openvpn3/config: Process --persist-tun during configuration import
If the configuration file being imported contains --persist-tun, it has
been ignored until now.  This option is not handled by the core VPN
client objects, as it expects this feature to be handled outside the
imported configuration profile.  This is by design, as this option is
considered to be platform specific; this might not be applicable on all
platforms.

Since Linux can handle this option, we will set the persist_tun property
to true if --persist-tun has been found in the configuration file.

This commit also introduces a simple test program which uses the
openvpn::ProfileMerge and openvpn::OptionList objects to process, parse
and query a configuration file for specific options.

Signed-off-by: David Sommerseth <davids@openvpn.net>