client: Fix issues with --persist-tun on single-use config profiles
When a configuration profile is tagged with single_use, the profile will
be automatically deleted by the configuration manager after the first
Fetch/GetConfig() method call.  When trying to retrieve attributes
*after* this call on such configuration, it will end up in a DBus
exception.

This patch moves retrieving the persist_tun flag to a point before the
Fetch/GetConfig() method call occurs.

Signed-off-by: David Sommerseth <davids@openvpn.net>