dbus/proxy: Add Ping() method in DBusProxy
This can be called as early as possible to ensure the service being
targeted manages to be auto-started by the D-Bus daemon.  On some
systems this auto-start can take a bit longer, which often makes the
calling process fail.

When calling Ping() it will also try 3 times with a 1 second delay
between each attempt.  This is to ensure auto-started services gets a
chance to settle.

Signed-off-by: David Sommerseth <davids@openvpn.net>