dbus: Introduce PropertyCollection for DBus objects
This makes it easier to add properties in D-Bus objects more dynamically
and it takes care of all the creation of introspection XML and ensuring
the return value needed when setting property values are correct.

The patch was originally written by Antonio and David restyled it and
made it do pointer based variable binding instead of storing values
inside the PropertyType class.

One challenge with the current implementation is that it does not handle
std::string and similar types of objects nicely; only primitive data
types which can be bound via a direct pointer are handled correctly.
But since this is a step forward in the right direction, this can be
fixed later on.

Signed-off-by: Antonio Quartulli <antonio@openvpn.net>
Signed-off-by: David Sommerseth <davids@openvpn.net>