cli/openvpn2: Fix python 2/3 compatibility - raw_input()
Python 3 does not have the raw_input() to get user input from the
console, it has been renamed to input().  But in Python2, input() means
something different - and is a feature not directly available in Python 3.

This adds a wrapper function, which calls raw_input() if Python 2.x or
older is detected, otherwise input() is used.

Signed-off-by: David Sommerseth <davids@openvpn.net>