sessionmgr: Catch exceptions if backend client call fails
In commit a209df737a0b583f63 it was implemented an improved update
mechanism of he backend status with the a status field in the session
object.  If his call threw an exception, the session object would exit
in anincorrect manner.  This situation with the backend is handled
elsewhere.  So lets just ignore the exception in this case and carry
on.

Signed-off-by: David Sommerseth <davids@openvpn.net>