configmgr: Fix wrong default log level
After the implementation of LogFilter in commit 275ae5abdb897, most of
the log events from the configuration manager disappeard.  The core of
the issue was that the default log level was set to be 3, while most log
events from this process happens on higher levels.

This implements setting the LogFilter log level via ConfigManagerSignals
in a more consistent way and paves the way for the main program to
provide the log level via an option or other configuration.

This change also improves some log messages and makes the lower level
messages a bit more user friendly.

Signed-off-by: David Sommerseth <davids@openvpn.net>