sessionmgr: Fix wrong default log level
The manager part of the session manager lost its capabilities to send
log events due to commit 275ae5abdb897.  This is a similar situation to
what was fixed in the configuration manager in commit 0027a707d7c2.

The session manager is slightly different in how it handles log events,
as each VPN client session may have different log levels.  This means
we have two set of log levels, manager log levels and session log
levels.  When a new SessionObject is established, it uses the manager
log level until the VPN client backend has reported back, then it
switches to a default log level for session objects.  This can further
be adjusted afterwards by manipulating the log_verbosity property in the
D-Bus session object.

The general logging has also been improved a bit, adding more log
messages where it is useful for the user and a sys-admin to know what is
happening.

Signed-off-by: David Sommerseth <davids@openvpn.net>