Improve the exit log messages of processes
Add a more explicit message when the process exits due to the idle exit
timer kicking in, as well as providing more information about the
process name both when this happens and when the signal handler is
acting.  This improves the logs, as this kind of log entries goes
outside of the D-Bus logs.

Also add this information only in programs defining the
SHUTDOWN_NOTIF_PROCESS_NAME macro.  This avoids printing noise on
applications not intended to run in the background as daemons.

Signed-off-by: David Sommerseth <davids@openvpn.net>