Replace custom C++14 check with upstream default C++ checks
Use C++14 when possible, otherwise ensure C++11 support is present.

This commit is slightly modified by David Sommerseth, updating
README.md and removing the explicit packaging of
ax_cxx_compile_stdcxx.m4, which is already packaged in reasonably newer
autoconf-archive packages.  This package exists in the oldest supported
Linux distribution, RHEL 7, so we can expect this package to be
available.

Signed-off-by: Arne Schwabe <arne@openvpn.net>
Signed-off-by: David Sommerseth <davids@openvpn.net>