lookup: Improve function declarations
lookup_username(), lookup_uid() and get_userid() where not ideally
declared.  So make all of them static, which is appropriate for pure
functions in header files.

And make get_userid() inline in addition.  This is such a simple
function where it causes no harm plus it silences a compiler warning
about unused functions; as this function may not always be used.

Signed-off-by: David Sommerseth <davids@openvpn.net>