python: Rename configparser and implement own error handler
Rename configparser.py to ConfigParser.py to correlate to the other
files in the openvpn3 module.

Also added a separate error handler which does not call print the usage
and call sys.exit().  This makes it easier to implement this module into
other programs and let those programs handle the exit phase.

Signed-off-by: David Sommerseth <davids@openvpn.net>