python: Implement UserInputSlot for Session user input interaction
Whenever a user input queue element/request is fetched, it will now be a
UserInputSlot object per backend request.  So if username and password
is requested, that will result in two UserInputSlot objects.

This new object will also provide a simpler API to the request,
including responding to the request.

Signed-off-by: David Sommerseth <davids@openvpn.net>