python: Added openvpn3-autoload
This is a script which is intended to be run during system boot-up,
where it will load all configurations from a pre-defined directory.
Configuration files in this directory must carry either the .conf
or .ovpn file extension in addition to be accompanied with a .autoload
configuration file.  This file is a JSON formatted file and provides
additional details not possible or suitable to put into the VPN
configuration profile.

Signed-off-by: David Sommerseth <davids@openvpn.net>