sessionmgr: Graceful handling when backend is not ready
If calling session object methods or retrieving properties requiring
interaction with the backend VPN process soon after the NewTunnel()
method has been called, it would most commonly crash the session
manager when the proxied D-Bus call to the backend failed.

This change makes the session manager more robust by actually checking
if the backend VPN process has registered to the session manager.  If
not, this can be reported back as an D-Bus exception.

Signed-off-by: David Sommerseth <davids@openvpn.net>