dbus: Add method for unique D-Bus bus name lookup
This returns the unique D-Bus bus name (:1.xxx) for a well-known D-Bus bus
name (like net.openvpn.v3.sessions).  This is useful when wanting to
verify if a D-Bus client call is coming from an expected service or not,
since the unique bus name is unpredictable while the well-known bus
name is static and specific to the service.  This look up is needed
since all the callback functions will provide the unique bus name of the
caller and not the well-known bus name.

Signed-off-by: David Sommerseth <davids@openvpn.net>