dbus: Add improved targeted signal sending
When sending D-Bus signals they can be unicast or broadcast, depending
on if a unique bus name is provided or not.  In some cases this is
either too broad or too limited, so this implements a more targeted
multicast.

This implementation is not ideal if enlisting too many recipients, but
for smaller sets of recipients this can be a simpler approach.

This feature can also be used to restrict signals from a full broadcast
to uni- or multicast later on during runtime.

Signed-off-by: David Sommerseth <davids@openvpn.net>