openvpn2/python: Require Python 3.x
Python 2.x is reaching EOL [1] January 1, 2020.  This is not in too far
future.  When it in addition adds code complexity to support Python 2.x
and 3.x in parallel, this makes even less sense.  So we are now removing
Python 2.x support completely.

This patch also fixes some odd import orders and adds a missing NL at
the end of the file.

[1] <https://www.python.org/dev/peps/pep-0373/>

Signed-off-by: David Sommerseth <davids@openvpn.net>