client: Add extra debug options
When debugging openvpn3-service-client via gdb, it can sometimes be
eaiser if the VPN client process don't run through the light
daemonization process (setsid() + fork()).

Here a couple of additional arguments are added if DEBUG_OPTIONS is
defined.  One which removes the setsid() call (--no-setsid) and one
for not doing the fork() (--no-fork).

Signed-off-by: David Sommerseth <davids@openvpn.net>