dbus/proxy: Use the GDBusConnection pointer in some proxy objects
When instantiating the DBusProxy object via OpenVPN3ConfigurationProxy,
DBusRequiresQueueProxy or OpenVPN3SessionProxy (via
DBusRequiresQueueProxy), use the GDBusConnection pointer directly
instead of the DBus object.  This will avoid the DBus object to close
D-Bus connection if more of these proxies shares the same D-Bus
connection when exiting its local scope, due to their destructor being
called - which drags in the DBus destructor.

This will need further improvements, but that will be much more invasive
and this solves an issue right now.

Signed-off-by: David Sommerseth <davids@openvpn.net>