sessionmgr: Improve behaviour with unregistered backends
If the backend VPN client process does not register within the session
manager, the session manager would easily segfault or misbehave in
various ways.

The changes added here ensures we always check if the backend has
registered before trying to do any calls towads it.  If the backend has
not registered, we do not have a valid proxy object to use for the
backend communication.

Further also allow a front-end to call the Disconnect method on an
unregistered session.  This is mostly to clean up sessions not being in
use.

Signed-off-by: David Sommerseth <davids@openvpn.net>