log: Implement a better generic log interface
This is a better implementation of a generic logging interface (API)
which can be used by most of the code, where the code generating the log
do not need to care where the log destination is.  This is handled via a
more proper C++ class abstraction.

Code wanting to generate log data need to implement the LogWriter class.
The code which is setting up the logging destinations can choose between
StreamLogWriter (text/plain on std::ostream), ColourStreamWriter
(similar to StreamLogWriter, with colours) and SyslogWriter (using
syslog).  The streams used by StreamLogWriter and ColourStreamWriter
needs to be opened before being passed to these log objects.

A unit test which tests all these three log destinations can be found
under ./src/tests/misc/logwriter-tests.cpp

Signed-off-by: David Sommerseth <davids@openvpn.net>