log: Implement --syslog in openvpn3-service-logger
This adds a --syslog argument which will the send all log events to
syslog instead of stdout or to a file.  We explicitly forbid combining
this with --colour, as such features should be handled by the syslog
service itself and not openvpn3-service-logger potentially corrupting
the syslog files with colour codes.

Signed-off-by: David Sommerseth <davids@openvpn.net>