log: Improve the log tag reference
The prioer log tag format looked like this:

    [:1.2625/net.openvpn.v3.backends]::

This format makes quick greps or searches a tad harder, as the search
expression needs to account for a few specific patterns.  In addition,
much of this information is duplicated in the metadata log line too.

There is a value to present a tag on each log line, as it makes it
easier to correlate log lines with each other - because each log
subscription kept by openvpn3-service-logger carries a unique tag.

Instead, use the std::hash representation of the tag above, and format
it in a more distitinctive way.  The new format looks like:

    {tag:15654764848252331829}

Signed-off-by: David Sommerseth <davids@openvpn.net>