client: Implement file and console logging
By adding --log-file to the command line when starting
openvpn3-service-client, it will write log data to the given filename in
addition to the D-Bus logging.  If filename is "stdout:", it will write
log data to stdout.  And for colourful log lines, add --colour.

It is by design it will only support logging to file or stdout.

Signed-off-by: David Sommerseth <davids@openvpn.net>