log: Added new method - LogWriter::WritePrepend()
This allows putting some data in front of the real log lines, similar to
AddMeta().  This also carries a flag which will use the same prepended
string on the meta log line as well.  This is useful when wanting to tag
lines in a consistent way across meta lines and log lines.

Signed-off-by: David Sommerseth <davids@openvpn.net>