log: SyslogWriter cannot change the timestamp flag
When using syslog, we also just presume that timestamps will always be
enabled.  So if anyone tries to change this flag, it will not be used.

Further, the SyslogWriter::Write() will also not care about this
internal timestamp flag, for the same reasoning.  We don't provide a
timestamp to the log lines we send via syslog().

Signed-off-by: David Sommerseth <davids@openvpn.net>