log: Make Attach() call syncrhonous
Make the LogServiceProxy::Attach() method call to the D-Bus service
syncrhonous.  This is to ensure the call was really processed by the
log service.  The (always empty) response is also ensured to be
unreferenced and freed, to avoid a memleak.

Further, also get rid of the additional 200µs sleep, as that should be
covered by the Ping() call.

Signed-off-by: David Sommerseth <davids@openvpn.net>