Extend struct StatusEvent to own conversion tasks itself
This adds a few more constructors which allows a quick setup of new
status events.  Also extend the GVariant parsing to cover tuples based
input "(uus)" in addition to the dictionary parsing.  On top of that,
add methods which generates both the dictionary and tuple based GVariant
objects based on its content.

StatusEvent also got extended with operator==() and operator!=(), to
more simply compare StatusEvent object.

And a fairly comprehensive unit test is also added.

Signed-off-by: David Sommerseth <davids@openvpn.net>