Revamp the status event processing
Make use of the new StatusEvent class, where the status event specific
parsing and processing is handled in that class.  This makes the overall
use and code handling simpler.

The code also had two different status event declarations in the D-Bus
implementation, both tuples and dictionary based.  Simplify that to only
use the tuples based variant.  It is not expected that these status
events will grow further.  And if it does, this will primarily be
handled in the StatusEvent class now.

Signed-off-by: David Sommerseth <davids@openvpn.net>