dbus/log: Provide a function to get the LogWriter object
This is needed to pass the LogWriter object to child-objects who
initialize their own signaling/logging objects but wants to re-use the
same logging mechanism as the rest of the program.

The alternative would be to save the LogWriter pointer in addition in
the parent object.  But since we already have this pointer in memory, we
just needed access to it.

Signed-off-by: David Sommerseth <davids@openvpn.net>