netcfg: Basic framework for the net.openvpn.v3.netcfg service
This service will be responsible for managing and configuring virtual
interfaces used by the OpenVPN 3 Client.  This will by default drop to
as few privileges and capabilties as needed at startup.

Currently, this only provides the basic D-Bus infrastructure and will
not do anything clever on the system; this needs to be implemented on
top of this.

Signed-off-by: David Sommerseth <davids@openvpn.net>