netcfg: Extend ResolverSettings to grasp GVariant
The D-Bus interface in netcfg uses an array of strings for modifying the
the DNS server and search domain lists.  This is received as a GVariant
object of the (as) {array of strings}.

This extends the ResolverSettings class to add support parsing GVariant
objects of the (as) type directly.  These four new methods are only
enabled if the glib2/gio headers have been included at an earlier point,
which enables code not requiring D-Bus to still use this class.

Signed-off-by: David Sommerseth <davids@openvpn.net>