netcfg: Add simple reference counting for ResolverSettings
This is to avoid a clash if multiple managed devices are in use and one
of the wants to disable the shared DNS settings.  Each managed devices
should always keep its own list over DNS entries it has added and remove
them independently and apply those settings if wanting to remove its own
settings.

Signed-off-by: David Sommerseth <davids@openvpn.net>