netcfg: Implement initial file based resolv.conf support
This is the last piece needed to implement direct manipulation of
/etc/resolv.conf via the netcfg service.

A new argument has been added, --resolv-conf, which is used to identify
which resolv.conf file to modify.  This is to enable support for various
file based DNS resolver environments, as well as enabling easier testing
and debugging.

When using this approach, the openvpn3-service-netcfg process also adds
the CAP_FOWNER capability, to be able to manipulate files and
directories the openvpn user account normally does not have access to.
This is a crude and broad covering privilege, but still better than
running as the root user.

Signed-off-by: David Sommerseth <davids@openvpn.net>