configmgr: Add config profile override support
This extends the configuration manager to support most of the various
override or site specific configuration parameters.  Many of them cannot
be provided by any configuration file option and can only be set
after importing a configuration profile.

This is slightly modified by David S, reorganizing the code slightly and
renaming some of the variables, functions and structs.

Signed-off-by: Arne Schwabe <arne@openvpn.net>
Signed-off-by: David Sommerseth <davids@openvpn.net>