ovpn3cli: Clean-up config-show
Move the displaying of configuration profile metadata to
config-manage --show instead of in the config-show command.  The latter
will now just dump the pure configuration profile and nothing else.

This commit has been slightly modified by David S. during the commit to
align it with the other adjustments done earlier as well as code style
cleanup.  Also removed the separator lines before and after the
configuration profile dump, which is no longer needed.

Signed-off-by: Arne Schwabe <arne@openvpn.net>
Signed-off-by: David Sommerseth <davids@openvpn.net>