sessionmgr: Add fine grained log access to sessions
The session manager only allows the session owner to change properties
in a D-Bus session object.  If the owner grants access to other users,
they would not get access to the session logs at all due to not being
able to change the receive_log_events and log_verbosity properties.

This adds a new boolean flag, restrict_log_access, which is true by
default.  When this is true, there is no change in behaviour; only the
owner may access the logs.  If this is changed to false, all users
granted access will be able to both manipulate the log_verbosity setting
and subscribe to the log events.

Signed-off-by: David Sommerseth <davids@openvpn.net>