configmgr: Improve reporting and avoid exceptions
There were a few cases where the D-Bus called could trigger a
DBusException to be thrown when changing override settings, which would
stop the openvpn3-service-configmr process - due to these exceptions
not being caught and processed properly.

This ensures a proper D-Bus error response is sent back to caller
instead of stopping the configmgr process.  It also improves the
error handling in openvpn3 config-manage to report situation back to
the front-end user

Signed-off-by: David Sommerseth <davids@openvpn.net>