configmgr: Improve data type checking in overrides
The configmgr proxy had an incorrect data type checking, which would
easily confuse if these errors occurred.  In addition the data type
checking inside the configmgr was poor, as it would be possible to pass
invalid data types if circumventing or abusing the configmgr proxy.

This just corrects and tightents the data type checking when modifying
configuration overrides.

Signed-off-by: David Sommerseth <davids@openvpn.net>