dbus/proxy: Add method to extract service version
Some of the OpenVPN 3 D-Bus services exposes a 'version' property.  The
GetServiceVersion() method will extract this property.

It does also carry a retry mechanism if the service is not found
instantly; which can happen if the service is in the
start-up/initialization phase and has not managed to establish the main
management object.

Signed-off-by: David Sommerseth <davids@openvpn.net>