configmgr/sessionmgr/proxy: Ensure the service has started
When the proxy objects for the Configuration and Session Managers is
being set up, do a call against the corresponding backend D-Bus service
to ensure it is available.  This help avoiding confusing "Unknown
method" errors if the service has not managed to settle before the real
D-Bus method calls comes.

Signed-off-by: David Sommerseth <davids@openvpn.net>