configmgr: Enforce the D-Bus type to be unit64_t for std::time_t
The data type timestamps in a D-Bus configuration profile changed from
uint64_t (t) to int64_t (x) due to the PropertyType improvmenent in
commit e1b1f5e3076ed8dfc9a, where the D-Bus data type is extracted from
the C++ data type.  This is due to std::time_t being handled as int64_t
on some systems.

As the prior implementation used uint64_t as the data type in the D-Bus
interface, keep it this way by overriding the D-Bus data type and force
it to be uint64_t, regardless of the C++ data type.

Signed-off-by: David Sommerseth <davids@openvpn.net>