dbus/proxy: Only fetch the service version on management objects
When the Configuration and Session Manager proxy objects are
initialized, they attempt to read the 'version' property.  This property
is only available in the main management objects, so there is no need to
try to fetch this on other objects.

Normally this would not be a probelm, as GetServiceVersion() will return
an empty string on objects not having this version property.  But ...

... The Session manager has a copy of the configuration path in the
session object.  So the openvpn3 sessions-list command will not attempt to
go via the main configuration object first when trying to extract the
configuration filename.  And the configuration manager can have this
object removed, which confuses the configuration proxy, as the error
returned if an object does not exist or if the D-Bus service has not
started or initialized yet are the same.  So the proxy code believes the
service is unavailable, and starts the delay loop to wait for it to settle
while the service is alive and available.

So by only calling the GetServiceVersion() method only when the object
path is targeting the main management object, this confusion can be
avoided and nothing breaks.

Signed-off-by: David Sommerseth <davids@openvpn.net>