client: Add client-side netcfg implementation
This is the first step towards the complete netcfg implementation on the
client side.  This establishes a new NetCfgTunBuilder class which is the
extension of the OpenVPN3 Core tunbuilder for configuring the tun
device.

Basically the NetCfgTunBuilder object will issue D-Bus calls towards
the net.openvpn.v3.netcfg service to get the TUN interface created and
configured.  In return a file descriptor is sent back to use for reading
and writing to the TUN device.

Signed-off-by: Arne Schwabe <arne@openvpn.net>