dbus/proxy: Add CheckObjectExists() method
Simplistic method to check if a D-Bus object really exists or not.  This
is to avoid calling Ping() to achieve a similar result.

The advantage of this new approach is that the result is returned much
quicker, due to the Ping() method will account for the target service
not running and will wait for it to potentially start up and settle.

Signed-off-by: David Sommerseth <davids@openvpn.net>