sessiongmr: Improve property query speed on session objects
The SessionObject would call the Ping() method on the backend client
service to check if the backend is available.  This is time consuming if
the backend has stopped running, as Ping() will account for the service
auto-start feature.

Instead, for these properties rather use CheckObjecctExists() instead.
This will quickly return with a status if the object exists or not,
which will result in a much quicker repsonse from the session manager.

Signed-off-by: David Sommerseth <davids@openvpn.net>