ovpn3cli: Use CheckObjectExists() where the path is known
In the openvpn3 command line, it would use Ping() to check if a user
provided D-Bus path exists.  As this can be slow, we rather use the new
CheckObjectExists() method in places where the user must know a D-Bus
path in advance.  This gives a better user experience by providing a
clearer error message when the path is incorrect or the overall slowness
is gone if the service was not running (in which the object does not
exist anyhow).

Signed-off-by: David Sommerseth <davids@openvpn.net>